# apps/server/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: EchoNote API
  version: 0.5.0
  description: EchoNote modular-monolith HTTP API.
paths:
...
  /api/v1/imports:
    post:
      operationId: createImport
      summary: Queue an episode import
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateImportRequest"
      responses:
        "202":
          description: Import queued.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/imports/{importId}:
    get:
      operationId: getImport
      summary: Get an import result
      parameters:
        - name: importId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Import state and resolved Episode when available.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/episodes:
    get:
      operationId: listEpisodes
      summary: List recently imported episodes
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 2147483647
            default: 0
      responses:
        "200":
          description: User-scoped Episode page.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EpisodeListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/episodes/{episodeId}:
    parameters:
      - name: episodeId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getEpisode
      summary: Get an Episode with Podcast and Sources
      responses:
        "200":
          description: Episode detail.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EpisodeDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteEpisode
      summary: Permanently delete an Episode
      responses:
        "204":
          description: Episode deleted.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/captures:
    post:
      operationId: createCapture
      summary: Capture a note for an Episode or URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaptureRequest"
      responses:
        "200":
          description: Idempotent replay of an existing capture.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureResponse"
        "201":
          description: Capture created; URL captures also queue an Import.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/episodes/{episodeId}/notes:
    parameters:
      - name: episodeId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: listEpisodeNotes
      summary: List active Notes for an Episode
      responses:
        "200":
          description: Notes ordered by client creation time, newest first.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createEpisodeNote
      summary: Create an idempotent Note for an Episode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "200":
          description: Idempotent replay of an existing Note.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "201":
          description: Note created.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/notes/{noteId}:
    parameters:
      - name: noteId
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateNote
      summary: Update a Note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteRequest"
      responses:
        "200":
          description: Note updated.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteNote
      summary: Soft-delete a Note
      responses:
        "204":
          description: Note deleted or already deleted.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  headers:
    RequestID:
      description: Request correlation identifier.
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Internal server error.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    CreateImportRequest:
      type: object
      additionalProperties: false
      required:
        - url
      properties:
        url:
          type: string
          minLength: 1
          maxLength: 4096
    ImportResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - url
        - status
        - stage
        - created_at
        - updated_at
      properties:
        id:
          type: string
        url:
          type: string
        status:
          type: string
          enum:
            - queued
            - running
            - succeeded
            - failed
            - canceled
        stage:
          type: string
        episode_id:
          type: string
        error:
          $ref: "#/components/schemas/ImportError"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ImportError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    CreateCaptureRequest:
      type: object
      additionalProperties: false
      description: Exactly one of episode_id and episode_url must be supplied.
      required:
        - client_note_id
        - content
        - created_at
      properties:
        client_note_id:
          type: string
        episode_id:
          type: string
        episode_url:
          type: string
          minLength: 1
          maxLength: 4096
        content:
          type: string
          minLength: 1
        created_at:
          type: string
          format: date-time
    CreateNoteRequest:
      type: object
      additionalProperties: false
      required:
        - client_note_id
        - content
        - created_at
      properties:
        client_note_id:
          type: string
        content:
          type: string
          minLength: 1
        created_at:
          type: string
          format: date-time
    UpdateNoteRequest:
      type: object
      additionalProperties: false
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
    Note:
      type: object
      additionalProperties: false
      required:
        - id
        - episode_id
        - client_note_id
        - content
        - created_at
        - updated_at
      properties:
        id:
          type: string
        episode_id:
          type: string
        client_note_id:
          type: string
        content:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
    CaptureResponse:
      type: object
      additionalProperties: false
      required:
        - note
      properties:
        note:
          $ref: "#/components/schemas/Note"
        import_id:
          type: string
    NoteListResponse:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Note"
...
    ResolveStatus:
      type: string
      enum:
        - pending
        - completed
        - failed
...
    PodcastSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - author
        - description
        - cover_url
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        description:
          type: string
        cover_url:
          type: string
        feed_url:
          type: string
    EpisodeSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - duration_ms
        - cover_url
        - resolve_status
        - transcription_status
        - ai_status
        - source_count
        - note_count
        - created_at
        - updated_at
      properties:
        id:
          type: string
        podcast:
          $ref: "#/components/schemas/PodcastSummary"
        title:
          type: string
        published_at:
          type: string
          format: date-time
        duration_ms:
          type: integer
          format: int64
          minimum: 0
        cover_url:
          type: string
        resolve_status:
          $ref: "#/components/schemas/ResolveStatus"
        transcription_status:
          $ref: "#/components/schemas/ProcessingStatus"
        ai_status:
          $ref: "#/components/schemas/ProcessingStatus"
        source_count:
          type: integer
          format: int64
          minimum: 0
        note_count:
          type: integer
          format: int64
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    EpisodeDetail:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - description
        - duration_ms
        - cover_url
        - resolve_status
        - transcription_status
        - ai_status
        - source_count
        - sources
        - created_at
        - updated_at
      properties:
        id:
          type: string
        podcast:
          $ref: "#/components/schemas/PodcastSummary"
        title:
          type: string
        description:
          type: string
        published_at:
          type: string
          format: date-time
        duration_ms:
          type: integer
          format: int64
          minimum: 0
        cover_url:
          type: string
        resolve_status:
          $ref: "#/components/schemas/ResolveStatus"
        transcription_status:
          $ref: "#/components/schemas/ProcessingStatus"
        ai_status:
          $ref: "#/components/schemas/ProcessingStatus"
        source_count:
          type: integer
          format: int64
          minimum: 0
        sources:
          type: array
          items:
            $ref: "#/components/schemas/EpisodeSource"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EpisodeListResponse:
      type: object
      additionalProperties: false
      required:
        - items
        - total
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EpisodeSummary"
        total:
          type: integer
          format: int64
          minimum: 0
        limit:
          type: integer
        offset:
          type: integer
...